+++ mne/evoked.py
@@ -168,7 +168,7 @@ class Evoked(ProjMixin, ContainsMixin, UpdateChannelsMixin,
         s += ", time : [%f, %f]" % (self.times[0], self.times[-1])
         s += ", n_epochs : %d" % self.nave
         s += ", n_channels x n_times : %s x %s" % self.data.shape
-        s += " (~%s)" % (sizeof_fmt(self._size),)
+        s += ", ~%s" % (sizeof_fmt(self._size),)
         return "<Evoked  |  %s>" % s
 
     @property